@@ -19,63 +19,35 @@ class Response implements ResponseContract
     use Macroable;
     
     // 输出头信息列表
-    protected $headers = [];
-    
-    // 跨域
-    protected $isAllowOrigin = false;
-    
-    // 允许跨域域名
-    protected $allowOrigin = '*';
-    
-    // 是否允许后续请求携带认证信息（cookies
-    // 该值只能是 true, 否则不返回，设置为 true | false
-    protected $allowCredentials = false; 
-    
-    // 预检结果缓存时间,缓存
-    protected $maxAge = '';
-    
-    // 该次请求的请求方式
-    protected $allowMethods = 'GET,POST,PATCH,PUT,DELETE,OPTIONS';
-    
-    // 该次请求的自定义请求头字段
-    protected $allowHeaders = 'X-Requested-With,X_Requested_With,Content-Type';
-    
-    // js 允许获取的 header 字段
-    protected $exposeHeaders = 'Authorization,authenticated';
+    protected array $headers = [];
     
     /**
-     * 是否允许跨域域名
+     * Initializes a new Response
+     *
+     * @param array $headers
      */
-    public function withIsAllowOrigin($isAllowOrigin = false)
+    public function __construct(array $headers = []) 
     {
-        if ($isAllowOrigin) {
-            $this->isAllowOrigin = true;
-        } else {
-            $this->isAllowOrigin = false;
-        }
-        
-        return $this;
+        $this->headers = array_merge($this->headers, $headers);
     }
     
     /**
-     * 允许跨域域名
+     * 允许跨域域名 *
      */
-    public function withAllowOrigin($allowOrigin = '*')
+    public function withAllowOrigin(string $allowOrigin): self
     {
-        $this->allowOrigin = $allowOrigin;
-        
-        return $this;
+        return $this->withHeader('Access-Control-Allow-Origin', $allowOrigin);
     }
     
     /**
      * 允许后续请求携带认证信息
      */
-    public function withAllowCredentials($allowCredentials = false)
+    public function withAllowCredentials(bool $allowCredentials): self
     {
+        // 是否允许后续请求携带认证信息(cookies)
+        // 该值只能是 true, 否则不返回
         if ($allowCredentials) {
-            $this->allowCredentials = true;
-        } else {
-            $this->allowCredentials = false;
+            return $this->withHeader('Access-Control-Allow-Credentials', "true");
         }
         
         return $this;
@@ -84,39 +56,56 @@ class Response implements ResponseContract
     /**
      * 预检结果缓存时间
      */
-    public function withMaxAge($maxAge = '')
+    public function withMaxAge(string $maxAge): self
     {
-        $this->maxAge = $maxAge;
+        if (! empty($maxAge)) {
+            return $this->withHeader('Access-Control-Max-Age', $maxAge);
+        }
         
         return $this;
     }
     
     /**
      * 该次请求的请求方式
+     * GET,POST,PATCH,PUT,DELETE,OPTIONS
      */
-    public function withAllowMethods($allowMethods = false)
+    public function withAllowMethods(string $allowMethods): self
     {
-        $this->allowMethods = $allowMethods;
-        
-        return $this;
+        return $this->withHeader('Access-Control-Allow-Methods', $allowMethods);
     }
     
     /**
      * 该次请求的自定义请求头字段
+     * X-Requested-With,X_Requested_With,Content-Type
      */
-    public function withAllowHeaders($allowHeaders = false)
+    public function withAllowHeaders(string $allowHeaders): self
     {
-        $this->allowHeaders = $allowHeaders;
-        
-        return $this;
+        return $this->withHeader('Access-Control-Allow-Headers', $allowHeaders);
     }
     
     /**
      * 设置 js 允许获取的 header 字段
+     * Authorization,authenticated
      */
-    public function withExposeHeaders($exposeHeaders = false)
+    public function withExposeHeaders(string $exposeHeaders): self
     {
-        $this->exposeHeaders = $exposeHeaders;
+        return $this->withHeader('Access-Control-Expose-Headers', $exposeHeaders);
+    }
+    
+    /**
+     * 判断 hader
+     */
+    public function hasHeader(string $name): bool
+    {
+        return isset($this->headers[$name]);
+    }
+    
+    /**
+     * 设置 hader
+     */
+    public function withHeader(string $name, mixed $content): self
+    {
+        $this->headers[$name] = $content;
         
         return $this;
     }
@@ -124,17 +113,29 @@ class Response implements ResponseContract
     /**
      * 设置 haders
      */
-    public function withHeader($name, $content = null)
+    public function withHeaders(array $headers): self
     {
-        if (is_array($name)) {
-            foreach ($name as $key => $value) {
-                $this->withHeader($key, $value);
-            }
-        } else {
-            if (! empty($name)) {
-                $this->headers[$name] = $content;
-            }
+        foreach ($headers as $key => $value) {
+            $this->withHeader($key, $value);
         }
+        
+        return $this;
+    }
+    
+    /**
+     * 删除 hader
+     */
+    public function deleteHeader(string $name): void
+    {
+        unset($this->headers[$name]);
+    }
+    
+    /**
+     * 重置 hader
+     */
+    public function resetHeader(): self
+    {
+        $this->headers = [];
         
         return $this;
     }
@@ -142,169 +143,141 @@ class Response implements ResponseContract
     /**
      * 获取 haders
      */
-    public function getHeaders()
+    public function getHeaders(): array
     {
-        if ($this->isAllowOrigin == 1) {
-            return $this->headers;
-        }
-        
-        return [];
-    }
-    
-    /**
-     * 组合跨域 headers
-     */
-    public function mergeCorsHeaders()
-    {
-        $header = [];
-        $header['Access-Control-Allow-Origin']  = $this->allowOrigin;
-        $header['Access-Control-Allow-Headers'] = $this->allowHeaders;
-        $header['Access-Control-Expose-Headers'] = $this->exposeHeaders;
-        $header['Access-Control-Allow-Methods'] = $this->allowMethods;
-        
-        if ($this->allowCredentials === true) {
-            $header['Access-Control-Allow-Credentials'] = "true";
-        }
-        
-        if (! empty($this->maxAge)) {
-            $header['Access-Control-Max-Age'] = $this->maxAge;
-        }
-        
-        $this->withHeader($header);
-        
-        return $this;
+        return $this->headers;
     }
     
     /**
      * 输出成功响应
      *
-     * @param  int         $code
-     * @param  string|null $message
-     * @param  array|null  $data
-     * @param  array       $userHeader
+     * @param  string $message
+     * @param  mixed  $data
+     * @param  int    $code
+     * @param  array  $headers
      * @return \Illuminate\Http\Response|\Illuminate\Contracts\Routing\ResponseFactory
      */
     public function success(
-        $message = "", 
-        $data = [], 
-        $code = \ResponseCode::SUCCESS, 
-        $userHeader = []
-    ) {
-        return $this->json(true, $code, $message, $data, $userHeader);
+        string $message = "", 
+        mixed  $data    = [], 
+        int    $code    = \ResponseCode::SUCCESS, 
+        array  $headers = []
+    ): mixed {
+        return $this->json(true, $code, $message, $data, $headers);
     }
     
     /**
      * 输出失败响应
      *
-     * @param  string|null $message
-     * @param  int         $code
-     * @param  array|null  $data
-     * @param  array       $userHeader
+     * @param  string $message
+     * @param  int    $code
+     * @param  mixed  $data
+     * @param  array  $headers
      * @return \Illuminate\Http\Response|\Illuminate\Contracts\Routing\ResponseFactory
      */
     public function error(
-        $message = "", 
-        $code = \ResponseCode::ERROR, 
-        $data = [], 
-        $userHeader = []
-    ) {
-        return $this->json(false, $code, $message, $data, $userHeader);
+        string $message = "", 
+        int    $code    = \ResponseCode::ERROR, 
+        mixed  $data    = [], 
+        array  $headers = []
+    ): mixed {
+        return $this->json(false, $code, $message, $data, $headers);
     }
     
     /**
      * 输出响应
      *
-     * @param  boolen      $success
-     * @param  int         $code
-     * @param  string|null $message
-     * @param  array|null  $data
-     * @param  array       $userHeader
+     * @param  boolen $success
+     * @param  int    $code
+     * @param  string $message
+     * @param  mixed  $data
+     * @param  array  $headers
      * @return \Illuminate\Http\Response|\Illuminate\Contracts\Routing\ResponseFactory
      */
     public function json(
-        $success = true, 
-        $code = \ResponseCode::INVALID, 
-        $message = "", 
-        $data = [], 
-        $userHeader = []
-    ) {
+        bool   $success = true, 
+        int    $code    = \ResponseCode::INVALID, 
+        string $message = "", 
+        mixed  $data    = [], 
+        array  $headers = []
+    ): mixed {
         $result = [
             'success' => $success,
             'code'    => $code,
         ];
         
-        if ($message) {
+        if (! empty($message)) {
             $result['message'] = $message;
         }
-        if ($data) {
+        
+        if (! empty($data)) {
             $result['data'] = $data;
         }
         
         // 返回 JSON 
-        return $this->returnJson($result, $userHeader);
+        return $this->returnJson($result, $headers);
     }
     
     /**
      * 将数组以标准 json 格式返回
      * 
      * @param  array $data
-     * @param  array $userHeader
+     * @param  array $headers
      * @return \Illuminate\Http\Response|\Illuminate\Contracts\Routing\ResponseFactory
      */
-    public function returnJson(array $data, $userHeader = []) 
+    public function returnJson(array $data, array $headers = []): mixed
     {
         $contents = json_encode($data, JSON_UNESCAPED_SLASHES|JSON_PRETTY_PRINT|JSON_UNESCAPED_UNICODE);
         
-        return $this->returnJsonFromString($contents, $userHeader);
+        return $this->returnJsonFromString($contents, $headers);
     }
     
     /**
      * 将 json 字符串以标准 json 格式返回
      * 
-     * @param  string|null $contents
-     * @param  array       $userHeader
+     * @param  string $contents
+     * @param  array  $headers
      * @return \Illuminate\Http\Response|\Illuminate\Contracts\Routing\ResponseFactory
      */
-    public function returnJsonFromString($contents, $userHeader = []) 
+    public function returnJsonFromString(string $contents, array $headers = []): mixed
     {
         // 添加 json 输出相应
-        $header = array_merge($userHeader, [
+        $headers = array_merge($headers, [
             'Content-Type' => 'application/json; charset=utf-8',
         ]);
         
-        return $this->returnData($contents, $header);
+        return $this->returnData($contents, $headers);
     }
     
     /**
      * 返回字符
      * 
-     * @param  string|null   $contents
-     * @param  array         $userHeader
+     * @param  string $contents
+     * @param  array  $headers
      * @return \Illuminate\Http\Response|\Illuminate\Contracts\Routing\ResponseFactory
      */
-    public function returnString($contents, $userHeader = []) 
+    public function returnString(string $contents, array $headers = []): mixed
     {
         // 文件输出相应
-        $header = array_merge($userHeader, [
+        $headers = array_merge($headers, [
             'Content-Type' => 'text/html',
         ]);
         
-        return $this->returnData($contents, $header);
+        return $this->returnData($contents, $headers);
     }
     
     /**
      * 返回数据
      * 
-     * @param  string|null $contents
-     * @param  array       $userHeader
+     * @param  string $contents
+     * @param  array  $headers
      * @return \Illuminate\Http\Response|\Illuminate\Contracts\Routing\ResponseFactory
      */
-    public function returnData($contents, $userHeader = []) 
+    public function returnData(string $contents, array $headers = []): mixed
     {
-        $this->mergeCorsHeaders()->withHeader($userHeader);
-        $header = $this->getHeaders();
+        $headers = $this->withHeaders($headers)->getHeaders();
         
-        return response($contents, 200, $header);
+        return response($contents, 200, $headers);
     }
 
 }
