@@ -15,12 +15,19 @@ interface Response
     /*
      * 响应json输出
      * 
-     * @param   boolen      $success
-     * @param   int         $code
-     * @param   string|null $message
-     * @param   array|null  $data
-     * @return  string      json
+     * @param  boolen $success
+     * @param  int    $code
+     * @param  string $message
+     * @param  mixed  $data
+     * @param  array  $header
+     * @return \Illuminate\Http\Response|\Illuminate\Contracts\Routing\ResponseFactory
      */
-    public function json($success, $code, $message, $data);
+    public function json(
+        bool   $success = true, 
+        int    $code = \ResponseCode::INVALID, 
+        string $message = "", 
+        mixed  $data = [], 
+        array  $header = []
+    ): mixed;
 
 }
