@@ -18,24 +18,24 @@ trait ResponseJson
      * 返回成功 json
      */
     protected function success(
-        $message = null, 
-        $data = null, 
-        $header = [],
-        $code = 0
+        string $message, 
+        mixed  $data = [], 
+        array  $headers = [],
+        int    $code = 0
     ) {
-        return Response::json(true, $code, $message, $data, $header);
+        return Response::json(true, $code, $message, $data, $headers);
     }
     
     /**
      * 返回错误 json
      */
     protected function error(
-        $message = null, 
-        $code = 1, 
-        $data = [], 
-        $header = []
+        string $message, 
+        int    $code = 1, 
+        mixed  $data = [], 
+        array  $headers = []
     ) {
-        return Response::json(false, $code, $message, $data, $header);
+        return Response::json(false, $code, $message, $data, $headers);
     }
     
     /**
